d2fsam-web-app/src/main/webapp/app/entities/organisation-unit-group-set/delete/organisation-unit-group-set-delete-dialog.component.ts
import { Component } from '@angular/core';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { IOrganisationUnitGroupSet } from '../organisation-unit-group-set.model';
import { OrganisationUnitGroupSetService } from '../service/organisation-unit-group-set.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  templateUrl: './organisation-unit-group-set-delete-dialog.component.html',
})
export class OrganisationUnitGroupSetDeleteDialogComponent {
  organisationUnitGroupSet?: IOrganisationUnitGroupSet;

  constructor(protected organisationUnitGroupSetService: OrganisationUnitGroupSetService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.organisationUnitGroupSetService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
